+++ service/src/main/kotlin/io/provenance/explorer/domain/exceptions/TendermintApiException.kt
@@ -4,3 +4,5 @@ class TendermintApiException(message: String) : Exception(message)
 
 class TendermintApiCustomException(val method: String, val httpMethod: String, val url: String, val status: Int, val body: String?) :
     Exception("$method: failed to $httpMethod $url: $status${body?.let { " body:$body" } ?: ""}")
+
+class FigmentApiException(message: String) : Exception(message)